Start support to Vega-Lite 4.14
The existing code is valid according to 4.14 (just missing the new
stuff).

Of note is the lack of support for optional types in the 0.11
release/ This would require a significant change which I don't
want to look at now, and there's also the fact that it's not so
clear the optional nature makes much difference for hvega use
cases.